chore: relocate response DTOs to 'res' directory
