+++ com.tech-nova.delivery/src/main/java/com/tech_nova/delivery/application/dto/res/DeliveryResponse.java
@@ -1,4 +1,4 @@
-package com.tech_nova.delivery.application.dto;
+package com.tech_nova.delivery.application.dto.res;
 
 import com.tech_nova.delivery.domain.model.delivery.Delivery;
 import com.tech_nova.delivery.domain.model.delivery.DeliveryStatus;
